# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IUT Cafeteria API
  description: |
    API for the IUT Cafeteria ordering system.

    ## Authentication
    All endpoints (except /login, /register, /health) require JWT authentication.
    Pass the token via:
    - `Authorization: Bearer <token>` header
    - `token` httpOnly cookie

    ## Correlation IDs
    All requests are assigned an `x-correlation-id` header for distributed tracing.
  version: 1.0.0
  contact:
    name: IUT Cafeteria Team
    email: admin@iut-cafeteria.app
servers:
  - url: http://localhost:3000
    description: Order Gateway (Development)
  - url: http://localhost:3001
    description: Identity Provider (Development)
  - url: http://localhost:3002
    description: Stock Service (Development)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: token
  schemas:
...
    MenuItem:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Chicken Biryani'
        description:
          type: string
          example: 'Fragrant rice with tender chicken'
        price:
          type: number
          format: float
          example: 120.00
        category:
          type: string
          enum: [rice, curry, snacks, drinks, dessert]
        is_available:
          type: boolean
        stock:
          type: integer
          example: 50
    Order:
      type: object
      properties:
        order_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING, STOCK_VERIFIED, IN_KITCHEN, READY, PICKED_UP, FAILED, CANCELLED]
        amount:
          type: number
          format: float
        item_name:
          type: string
        created_at:
          type: string
          format: date-time
...
  parameters:
...
    IdempotencyKey:
      name: x-idempotency-key
      in: header
      description: Idempotency key for safe retries
      schema:
        type: string
        format: uuid
paths:
...
  /menu:
    get:
      tags: [Menu]
      summary: Get available menu items
      security:
        - BearerAuth: []
        - CookieAuth: []
      responses:
        '200':
          description: List of menu items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/MenuItem'
...
  /orders:
    get:
      tags: [Orders]
      summary: Get order history
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Order history with pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer

    post:
      tags: [Orders]
      summary: Place a new order
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [itemId]
              properties:
                itemId:
                  type: integer
                type:
                  type: string
                  enum: [dinner, iftar]
                  default: dinner
                scheduledPickupTime:
                  type: string
                  format: date-time
      responses:
        '202':
          description: Order accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  orderId:
                    type: string
                    format: uuid
                  message:
                    type: string
                  item:
                    type: string
        '402':
          description: Insufficient balance
        '409':
          description: Out of stock
...
  /reviews:
    post:
      tags: [Orders]
      summary: Submit order review
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderId, rating]
              properties:
                orderId:
                  type: string
                  format: uuid
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                comment:
                  type: string
      responses:
        '200':
          description: Review submitted

  # ─── Admin ────────────────────────────────────────────────────
  /admin/menu:
    get:
      tags: [Admin]
      summary: List all menu items (including unavailable)
      responses:
        '200':
          description: Full menu list

    post:
      tags: [Admin]
      summary: Create menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, price, category]
              properties:
                name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                category:
                  type: string
                is_available:
                  type: boolean
                  default: true
                image_url:
                  type: string
                initial_stock:
                  type: integer
                  default: 50
      responses:
        '201':
          description: Item created
...